perf: send admin translations once and keep caches after install
- Share admin translations (~40KB) as an Inertia once prop keyed by locale
  and a translation revision, so admin navigation stops re-sending them on
  every visit; editing translations or switching locale refreshes them.
- Rebuild config/route/event caches after the installer finishes in
  production; it used to clear them and leave every request uncached.
- Fix a float-to-int deprecation in the dashboard uptime calculation.

+++ app/Http/Controllers/DashboardController.php
@@ -286,7 +286,7 @@ class DashboardController extends Controller
     private function getServerUptime(): string
     {
         $startTime = defined('LARAVEL_START') ? LARAVEL_START : time();
-        $uptimeSeconds = time() - $startTime;
+        $uptimeSeconds = (int) (time() - $startTime);
 
         $days = floor($uptimeSeconds / 86400);
         $hours = floor(($uptimeSeconds % 86400) / 3600);
